source cleanup in latest eclipse, using Pivot profile
git-svn-id: https://svn.apache.org/repos/asf/pivot/branches/2.0.x@1610977 13f79535-47bb-0310-9956-ffa450edef68

+++ wtk/src/org/apache/pivot/wtk/validation/IntRangeValidator.java
@@ -69,8 +69,8 @@ public class IntRangeValidator extends IntValidator {
         boolean valid = false;
 
         if (super.isValid(text)) {
-            BigInteger min = BigInteger.valueOf((long)minValue);
-            BigInteger max = BigInteger.valueOf((long)maxValue);
+            BigInteger min = BigInteger.valueOf(minValue);
+            BigInteger max = BigInteger.valueOf(maxValue);
             BigInteger value = new BigInteger(text);
             valid = value.compareTo(min) >= 0 && value.compareTo(max) <= 0;
         }
